pyproject.toml
[project]
name = "mcp-web-deploy"
version = "0.1.0"
description = "MCP server for libcloud operations with deploy_html tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.9.2",
    "apache-libcloud>=3.8.0",
    "pyyaml>=6.0.2",
    "loguru>=0.7.3",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
mcp-web-deploy = "main:run_server"
